Restructuring for handling obs. data and removing deprecated files.
