+++ .gitignore
@@ -1,7 +1,9 @@
 ### Project specific
 # ignore...
 data/cases/*
-data/input/*
+data/input/clm/*
+data/input/user_observations/atmosphere/*
+data/input/user_observations/soil/*
 data/output/*
 landsites_tools/custom_settings/*
 *.remote-sync.json
@@ -9,7 +11,10 @@ landsites_tools/custom_settings/*
 
 # keep... (don't track user files, but make sure folders are in place)
 !cases/.gitkeep
-!input/.gitkeep
+!data/input/clm/.gitkeep
+!data/input/user_observations/.gitkeep
+!data/input/user_observations/atmosphere/example.csv
+!data/input/user_observations/soil/example.csv
 !output/.gitkeep
 !landsites_tools/custom_settings/.gitkeep
