+++ landsites_tools/inputprep/create_new_input.py
@@ -1,3 +1,3 @@
 #!/usr/bin/env python
 
-"""Coming soon."""
+"""To be discussed."""
