+++ landsites_tools/inputprep/obs_adjust_input.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+
+"""Updates existing input data files to replace the atmospheric forcing from
+global data sets with local observations provided in a csv file. Applies gap-
+filling to missing values (if appropriate). WARNING: See docs for observational
+data requirements and formats!"""
