+++ old_scripts/lightning_site_clm5.ncl
