+++ old_scripts/prepare_atm_forcing_gswp3.ncl
